+++ Changelog.md
@@ -2,6 +2,8 @@
 
 ## Version (NEXT)
 * use a single state atom in simulator
+* added perfect power factor to math
+
 
 ## Version 0.1.0
 * implemented quantom states, quantum gates, gate registry and quantum cirquits